Extract ASN.1 DER
