+++ src/JWT.php
@@ -3,6 +3,7 @@
 namespace Firebase\JWT;
 
 use DateTime;
+use Firebase\JWT\ASN\DerConverter;
 use Firebase\JWT\Exceptions\DomainException;
 use Firebase\JWT\Exceptions\InvalidArgumentException;
 use Firebase\JWT\Exceptions\UnexpectedValueException;
@@ -25,10 +26,6 @@ use Firebase\JWT\Exceptions\ExpiredException;
  */
 class JWT
 {
-    const ASN1_INTEGER = 0x02;
-    const ASN1_SEQUENCE = 0x10;
-    const ASN1_BIT_STRING = 0x03;
-
     /**
      * When checking nbf, iat or expiration times,
      * we want to provide some extra leeway time to
@@ -107,7 +104,7 @@ class JWT
         }
         if ($header->alg === 'ES256') {
             // OpenSSL expects an ASN.1 DER sequence for ES256 signatures
-            $sig = self::signatureToDER($sig);
+            $sig = DerConverter::signatureToDER($sig);
         }
 
         if (\is_array($key) || $key instanceof \ArrayAccess) {
@@ -215,7 +212,7 @@ class JWT
                     throw new DomainException("OpenSSL unable to sign data");
                 } else {
                     if ($alg === 'ES256') {
-                        $signature = self::signatureFromDER($signature, 256);
+                        $signature = DerConverter::signatureFromDER($signature, 256);
                     }
                     return $signature;
                 }
@@ -392,125 +389,5 @@ class JWT
             return \mb_strlen($str, '8bit');
         }
         return \strlen($str);
-    }
-
-    /**
-     * Convert an ECDSA signature to an ASN.1 DER sequence
-     *
-     * @param   string $sig The ECDSA signature to convert
-     * @return  string The encoded DER object
-     */
-    private static function signatureToDER($sig)
-    {
-        // Separate the signature into r-value and s-value
-        list($r, $s) = \str_split($sig, (int) (\strlen($sig) / 2));
-
-        // Trim leading zeros
-        $r = \ltrim($r, "\x00");
-        $s = \ltrim($s, "\x00");
-
-        // Convert r-value and s-value from unsigned big-endian integers to
-        // signed two's complement
-        if (\ord($r[0]) > 0x7f) {
-            $r = "\x00" . $r;
-        }
-        if (\ord($s[0]) > 0x7f) {
-            $s = "\x00" . $s;
-        }
-
-        return self::encodeDER(
-            self::ASN1_SEQUENCE,
-            self::encodeDER(self::ASN1_INTEGER, $r) .
-            self::encodeDER(self::ASN1_INTEGER, $s)
-        );
-    }
-
-    /**
-     * Encodes a value into a DER object.
-     *
-     * @param   int     $type DER tag
-     * @param   string  $value the value to encode
-     * @return  string  the encoded object
-     */
-    private static function encodeDER($type, $value)
-    {
-        $tag_header = 0;
-        if ($type === self::ASN1_SEQUENCE) {
-            $tag_header |= 0x20;
-        }
-
-        // Type
-        $der = \chr($tag_header | $type);
-
-        // Length
-        $der .= \chr(\strlen($value));
-
-        return $der . $value;
-    }
-
-    /**
-     * Encodes signature from a DER object.
-     *
-     * @param   string  $der binary signature in DER format
-     * @param   int     $keySize the number of bits in the key
-     * @return  string  the signature
-     */
-    private static function signatureFromDER($der, $keySize)
-    {
-        // OpenSSL returns the ECDSA signatures as a binary ASN.1 DER SEQUENCE
-        list($offset, $_) = self::readDER($der);
-        list($offset, $r) = self::readDER($der, $offset);
-        list($offset, $s) = self::readDER($der, $offset);
-
-        // Convert r-value and s-value from signed two's compliment to unsigned
-        // big-endian integers
-        $r = \ltrim($r, "\x00");
-        $s = \ltrim($s, "\x00");
-
-        // Pad out r and s so that they are $keySize bits long
-        $r = \str_pad($r, $keySize / 8, "\x00", STR_PAD_LEFT);
-        $s = \str_pad($s, $keySize / 8, "\x00", STR_PAD_LEFT);
-
-        return $r . $s;
-    }
-
-    /**
-     * Reads binary DER-encoded data and decodes into a single object
-     *
-     * @param string $der the binary data in DER format
-     * @param int $offset the offset of the data stream containing the object
-     * to decode
-     * @return array [$offset, $data] the new offset and the decoded object
-     */
-    private static function readDER($der, $offset = 0)
-    {
-        $pos = $offset;
-        $size = \strlen($der);
-        $constructed = (\ord($der[$pos]) >> 5) & 0x01;
-        $type = \ord($der[$pos++]) & 0x1f;
-
-        // Length
-        $len = \ord($der[$pos++]);
-        if ($len & 0x80) {
-            $n = $len & 0x1f;
-            $len = 0;
-            while ($n-- && $pos < $size) {
-                $len = ($len << 8) | \ord($der[$pos++]);
-            }
-        }
-
-        // Value
-        if ($type == self::ASN1_BIT_STRING) {
-            $pos++; // Skip the first contents octet (padding indicator)
-            $data = \substr($der, $pos, $len - 1);
-            $pos += $len - 1;
-        } elseif (!$constructed) {
-            $data = \substr($der, $pos, $len);
-            $pos += $len;
-        } else {
-            $data = null;
-        }
-
-        return array($pos, $data);
     }
 }
